requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

# Core dependencies
azure-functions
requests
azure-eventhub

# Development dependencies (optional)
# pytest
# pytest-mock
# requests-mock
# pytest-cov
